Integrate WorkflowNodes to job.py
